Move all render code to Renderer class
