+++ alvr/server/cpp/platform/linux/FrameRender.h
@@ -1,117 +1,54 @@
 #pragma once
 
-#include <vector>
-#include <string>
-#include <vulkan/vulkan.h>
+#include "Renderer.h"
+#include "ffmpeg_helper.h"
+#include "protocol.h"
 
-class RenderPipeline;
-
-class FrameRender
+class FrameRender : public Renderer
 {
 public:
-    struct Output {
-        VkImage image;
-        VkImageCreateInfo imageInfo;
-        VkDeviceSize size;
-        VkDeviceMemory memory;
-        // ---
-        VkImageView view;
-        VkFramebuffer framebuffer;
-    };
-
-    explicit FrameRender(const VkInstance &inst, const VkDevice &dev, const VkPhysicalDevice &physDev, const std::vector<const char *> &devExtensions);
+    explicit FrameRender(alvr::VkContext &ctx, init_packet &init, int fds[]);
     ~FrameRender();
 
-    void Startup(uint32_t width, uint32_t height, VkFormat format, uint32_t queueFamilyIndex);
-
-    void AddImage(VkImageCreateInfo imageInfo, size_t memoryIndex, int imageFd, int semaphoreFd);
-
-    void AddPipeline(RenderPipeline *pipeline);
-
-    Output CreateOutput(uint32_t width, uint32_t height);
-
-    void Render(uint32_t index, uint64_t waitValue);
+    Output CreateOutput();
+    uint32_t GetEncodingWidth() const;
+    uint32_t GetEncodingHeight() const;
 
 private:
-    struct InputImage {
-        VkImage image;
-        VkDeviceMemory memory;
-        VkSemaphore semaphore;
-        VkImageView view;
-        VkDescriptorSet descriptor;
+    struct ColorCorrection {
+        float renderWidth;
+        float renderHeight;
+        float brightness;
+        float contrast;
+        float saturation;
+        float gamma;
+        float sharpening;
+        float _align;
     };
 
-    struct StagingImage {
-        VkImage image;
-        VkDeviceMemory memory;
-        VkImageView view;
-        VkFramebuffer framebuffer;
-        VkDescriptorSet descriptor;
+    struct FoveationVars {
+        int32_t targetEyeWidth;
+        int32_t targetEyeHeight;
+        int32_t optimizedEyeWidth;
+        int32_t optimizedEyeHeight;
+
+        float eyeWidthRatio;
+        float eyeHeightRatio;
+
+        float centerSizeX;
+        float centerSizeY;
+        float centerShiftX;
+        float centerShiftY;
+        float edgeRatioX;
+        float edgeRatioY;
     };
 
-    void commandBufferBegin();
-    void commandBufferSubmit();
-    void addStagingImage(uint32_t width, uint32_t height);
-    uint32_t memoryTypeIndex(VkMemoryPropertyFlags properties, uint32_t typeBits) const;
+    void setupColorCorrection();
+    void setupFoveatedRendering();
 
-    struct {
-        PFN_vkImportSemaphoreFdKHR vkImportSemaphoreFdKHR;
-        bool haveDrmModifiers = false;
-    } d;
-
-    Output m_output;
-    std::vector<InputImage> m_images;
-    std::vector<StagingImage> m_stagingImages;
+    uint32_t m_width;
+    uint32_t m_height;
+    ColorCorrection m_colorCorrectionPushConstants;
+    FoveationVars m_foveatedRenderingPushConstants;
     std::vector<RenderPipeline*> m_pipelines;
-
-    VkInstance m_inst;
-    VkDevice m_dev;
-    VkPhysicalDevice m_physDev;
-    VkQueue m_queue;
-    uint32_t m_queueFamilyIndex;
-    VkFormat m_format;
-    VkExtent2D m_imageSize;
-    VkCommandPool m_commandPool;
-    VkSampler m_sampler;
-    VkBuffer m_vertexBuffer;
-    VkDeviceMemory m_vertexMemory;
-    VkRenderPass m_renderPass;
-    VkDescriptorPool m_descriptorPool;
-    VkDescriptorSetLayout m_descriptorLayout;
-    VkCommandBuffer m_commandBuffer;
-    VkFence m_fence;
-
-    friend class RenderPipeline;
-};
-
-class RenderPipeline
-{
-public:
-    enum ShaderType {
-        VertexShader,
-        FragmentShader
-    };
-
-    explicit RenderPipeline(FrameRender *render);
-    ~RenderPipeline();
-
-    void SetShader(ShaderType type, const char *filename);
-    void SetShader(ShaderType type, const unsigned char *data, unsigned len);
-    void SetPushConstant(ShaderType type, const void *data, uint32_t size);
-
-private:
-    void Build();
-    void Render(VkDescriptorSet in, VkFramebuffer out, VkRect2D outSize);
-
-    FrameRender *r;
-    VkShaderModule m_vertexShader = VK_NULL_HANDLE;
-    VkShaderModule m_fragmentShader = VK_NULL_HANDLE;
-    const void *m_vertexConstant = nullptr;
-    uint32_t m_vertexConstantSize = 0;
-    const void *m_fragmentConstant = nullptr;
-    uint32_t m_fragmentConstantSize = 0;
-    VkPipeline m_pipeline = VK_NULL_HANDLE;
-    VkPipelineLayout m_pipelineLayout = VK_NULL_HANDLE;
-
-    friend class FrameRender;
 };
